ReadQueue
This code will allow us to detect collisions on the bus.  I tested with 0 and with 20_000 delay and it seems to be functioning correctly.  Will test fully once the delays are added in to the rest of the code.  For now `word_time` will be set to 0 for "immediate" transmission.